Issue 37 integration process (#39)
Introduce function to drive the process of introducing candidates/applicants to SAP.
At this point we've yet to get a successful introduction completed.  The POST is happening and we're getting back valid failure/validation REST responses.  We just need to figure out what business rules/formatting is causing the submissions to fail.