Attempt to deal better with 429 issues.  Vastly simplify candidate processing code, and a couple other small simplifications.
More on 429s
------------
What has been done is very low-tech:  just request fewer employees from SR for now.  This is really not sufficient, because if we have more than the limit (currently 15) of candidates, and >limit of those candidates are already introduced, there's a chance we will pick up only introduced candidates and thus skip any non-introduced ones in any given run.

The problem is, to find out if a candidate is introduced, we have to do the property GET from SR, so there's a bit of a chicken-and-egg issue trying to avoid the 429, unless we can throttle correctly.

All that said, this commit is just to get the Lambda running again without 429s.  As it is, that's all we ever see, so it's difficult to reason about any other issues that come up.